+++ core/domain/src/main/kotlin/com/nexters/bandalart/android/core/domain/usecase/bandalart/DeleteBandalartCellUseCase.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.core.domain.usecase
+package com.nexters.bandalart.android.core.domain.usecase.bandalart
 
 import com.nexters.bandalart.android.core.domain.repository.BandalartRepository
 import com.nexters.bandalart.android.core.domain.util.runSuspendCatching
@@ -6,12 +6,15 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class CreateBandalartUseCase @Inject constructor(
+class DeleteBandalartCellUseCase @Inject constructor(
   private val repository: BandalartRepository,
 ) {
-  suspend operator fun invoke() {
+  suspend operator fun invoke(
+    bandalartKey: String,
+    cellKey: String,
+  ) {
     runSuspendCatching {
-      repository.createBandalart()
+      repository.deleteBandalartCell(bandalartKey, cellKey)
     }
   }
 }
