+++ core/domain/src/main/kotlin/com/nexters/bandalart/android/core/domain/usecase/bandalart/UpdateBandalartCellUseCase.kt
@@ -1,20 +1,22 @@
-package com.nexters.bandalart.android.core.domain.usecase
+package com.nexters.bandalart.android.core.domain.usecase.bandalart
 
+import com.nexters.bandalart.android.core.domain.entity.UpdateBandalartCellEntity
 import com.nexters.bandalart.android.core.domain.repository.BandalartRepository
 import com.nexters.bandalart.android.core.domain.util.runSuspendCatching
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class DeleteBandalartCellUseCase @Inject constructor(
+class UpdateBandalartCellUseCase @Inject constructor(
   private val repository: BandalartRepository,
 ) {
   suspend operator fun invoke(
     bandalartKey: String,
     cellKey: String,
+    updateBandalartCellEntity: UpdateBandalartCellEntity,
   ) {
     runSuspendCatching {
-      repository.deleteBandalartCell(bandalartKey, cellKey)
+      repository.updateBandalartCell(bandalartKey, cellKey, updateBandalartCellEntity)
     }
   }
 }
